更新BenchSuite项目以支持ExamLab的新功能： 1. 添加对SetAnimationTiming知识点的支持 2. 改进幻灯片版式检测，支持新的11个标准版式 3. 优化页脚文字检测（已兼容删除的FooterTextValue参数） 4. 添加版式名称标准化和映射功能 5. 创建兼容性测试程序 6. 确保decimal分数类型的完全支持

+++ ExamLab/Models/ExamModule.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.ObjectModel;
 using System.Linq;
+using ExamLab.Services;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
-using ExamLab.Services;
 
 namespace ExamLab.Models;
 
@@ -110,7 +110,7 @@ public class ExamModule : ReactiveObject
     /// <summary>
     /// 总分
     /// </summary>
-    public double TotalScore => Questions.Sum(q => q.TotalScore);
+    public decimal TotalScore => Questions.Sum(q => q.TotalScore);
 
     /// <summary>
     /// 操作点数量
